#22 [오늘도 수고한나에게(운동 코멘트)추가] - pace 타입변경

+++ api/src/main/java/com/wancs/battle_run/domain/running/dto/TotalRecordInterface.java
@@ -4,5 +4,5 @@ public interface TotalRecordInterface {
     Float getTotalDistance();
     Long getTotalRunningTime();
     Integer getTotalCalorie();
-    Double getTotalPace();
+    Integer getTotalPace();
 }
